+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Token` type to be a struct containing a `std::borrow::Cow` for the token content instead of
   an enum of `RefToken` and `OwningToken`.
 - `--files-file` parameter to `--files-from`.
+- `FormattingData` fields from `usize` to `u16`.
 
 ### Fixed